attempt to convert base and btos2 rolelists when used in the other env
Some roles and buckets exist in base game but not btos2, and vice versa.

Adds a static class to convert those roles and buckets, as well as a static class to check if BTOS is loaded. (otherwise it will error out if the user does not have btos2 installed)

Also, increment the version to 1.0.1